Implement perspective notarization
Based on how we have things, I THINK all we needed to do here was keep a running list of all seen Signatures for the epoch's block and then overwrite the Message's current list with that list.

The notarization logic will take care of the check for valid signatures so besides some potential space concerns (which would only last for one epoch anyway), this seems alright to me.

Fixed minor bug with notarization extension index logging.